add weird binding.gyp which checks for icu-config
only when icu-config is available it actually tries to build
node-stringprep.cc.

i added support for windows and linux seperatly like pg (node-postgres)
does. might not be pretty but looks like that it works (at least on my
machine it builds with and without libicu-dev without errors).
Haven't tested it on windows though.